==> your-project/3. one-hot-encoding_country-genre.ipynb <==
# Libraries
import pandas as pd
import numpy as np
import re

netflix = pd.read_csv("../data/netflix_cleandata.csv", index_col=0)

# Getting the unique values
countries = list(netflix["country"].apply(str).unique())
countries = [x for x in countries if x != "nan"]

# Getting the first country from the unique

only_first_country = []

for country in countries:
    only_first_country.extend(re.findall(r"^(.+?),", country))
    
# Removing duplicates and sorting the countries
countries_clean = list(set(only_first_country))
countries_clean.sort()

def country_hot_encoded(df, country):
    
    """
    Input: Dataframe, list of unique countries
    
    Output: The function returns a SUM one-hot encoded of the countries
    """
    
    list_of = []
    for c in country:
        country_1 = df[df["country"].str.contains(c, case=False, regex=False, na=False)]
        df2 = country_1[["listed_in"]].unstack().str.get_dummies(sep=", ").sum(level = 0).rename(index={"listed_in":c})
        list_of.append(df2)
          
    result = pd.concat(list_of).fillna(0).astype(int)
    
    return result

country_hot_encoded(netflix, countries_clean)

# Storing the one-hot encoding into a variable
country_encoded = country_hot_encoded(netflix, countries_clean)

# Saving the file into a CSV
#country_encoded.to_csv("../data/country_hotencodedtest.csv", sep=";")

# Unstacking DF for Tableau importing
tableau_country = country_encoded.unstack().reset_index()

# Change column names for better understanding
tableau_country = tableau_country.rename(columns={"level_0":"genre", "level_1":"country", 0:"count"})

# Saving the file into a CSV
#tableau_country.to_csv("../data/genres_per_country.csv")
